Refactor: removed ONYXKEYS.COLLECTION.POLICY from PolicyUtils
